Fixed K9 scan issues.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@10781 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdePkg/Library/BaseLib/MultS64x64.c
@@ -38,5 +38,5 @@ MultS64x64 (
   IN      INT64                     Multiplier
   )
 {
-  return (INT64)MultU64x64 (Multiplicand, Multiplier);
+  return (INT64)MultU64x64 ((UINT64) Multiplicand, (UINT64) Multiplier);
 }
